added smsparser logic and blankPurchases proxy to API

+++ CostManager/CommonUtilities/Models/Purchase.cs
@@ -7,7 +7,7 @@ namespace CommonUtilities.Models
     {
         public List<Position> Positions { get; set; }
         public string Market { get; set; }
-        public int TotalSum { get; set; }
+        public double TotalSum { get; set; }
         public DateTime PurchaseTime { get; set; }
         public string Guid { get; set; }
         public string UserGuid { get; set; }
